Make compatible again with latest Carbon version, fixes #108
