+++ src/Lang/cs.php
@@ -45,11 +45,11 @@ return array(
     'saturday'  => 'sobota',
     'sunday'    => 'neděle',
 	
-    'year_ago'      => 'rokem|[2,Inf]:count lety',
+	'year_ago'      => 'rokem|[2,Inf]:count lety',
     'month_ago'     => 'měsícem|[2,Inf]:count měsíci',
     'week_ago'      => 'týdnem|[2,Inf]:count týdny',
     'day_ago'       => 'dnem|[2,Inf]:count dny',
     'hour_ago'      => 'hodinou|[2,Inf]:count hodinami',
-    'minute_ago'    => 'minutou|[2,Inf]:count minutami',
+    'minute_ago'    => 'minútou|[2,Inf]:count minutami',
     'second_ago'    => 'sekundou|[2,Inf]:count sekundami',
 );
