+++ src/Lang/sk.php
